+++ NEWS.md
@@ -12,3 +12,18 @@
 * Remove config and routes comment lines
 * Setup script
 * Setup secret token
+
+1.0.0 (06 Abril 2016)
+
+* Add `quiet_assets` as development dependency
+* Generate rake, rails and rspec binstubs with Spring
+* Make Shoulda Matchers work with Spring
+* Add `binding_of_caller` gem
+* Add `better_errors` gem
+* Add `pry-byebug` and `pry-rails` gems
+* Add `bullet` as development dependency
+* Add Bundler Audit to scan Gemfile for insecure dependencies per CVEs.
+* Add Dotenv to development and test environments to load environment variables from the `.env` file.
+* Add `brakeman`
+* Add `letter_opener`
+* Add `rubocop`